Fix code smells from sonarcloud

+++ src/view/block/list_directory.cc
@@ -175,7 +175,7 @@ bool ListDirectory::OnCustomEvent(const CustomEvent& event) {
       // To get a better experience, update focused and select indexes,
       // to highlight current playing song entry in list
       auto it = std::find(entries_.begin(), entries_.end(), *curr_playing_);
-      int index = it != entries_.end() ? it - entries_.begin() : 0;
+      int index = it != entries_.end() ? static_cast<int>(it - entries_.begin()) : 0;
 
       focused_ = index;
       selected_ = index;
@@ -211,7 +211,6 @@ bool ListDirectory::OnCustomEvent(const CustomEvent& event) {
     // In case that song has finished successfully, attempt to play next one
     if (auto content = event.GetContent<model::Song::CurrentInformation>();
         content.state == model::Song::MediaState::Finished) {
-
       if (auto file = SelectNextToPlay(); !file.empty()) {
         LOG("Song finished, attempt to play next file: ", file);
         // Send user action to controller
